comment out other main functions
In order to select which one to compile, uncomment each.  I might be able to stick each of these in seperate functions and shove it all in one file, depending on the marker's wishes on visibility.  I might also need to split these files back up into their own projects.  We'll see.